fix: duplicated validation

+++ src/Shared/Communication/Requests/RequestRegisterUser.cs
@@ -4,7 +4,7 @@ public class RequestRegisterUser
 {
 	public string Email { get; set; } = string.Empty;
 	public string Password { get; set; } = string.Empty;
-	public decimal Balance { get; set; } 
+	public int Balance { get; set; } 
 	public string FullName { get; set; } = string.Empty;
 	public string? Cpf { get; set; }
 	public string? Cnpj { get; set; }
